added more methods to handle error cases, restructured the api folder
