+++ lib/apis/error/api_error_helper.dart
@@ -1,5 +1,6 @@
-import '../extensions.dart';
-import '../mmd_ecommerce.dart';
+import 'package:graphql_flutter/graphql_flutter.dart';
+import 'package:mmd_ecommerce_fl_lib/apis/base/api_keys.dart';
+import 'package:mmd_ecommerce_fl_lib/extensions.dart';
 
 class ApiErrorHelper {
   static bool isValidationError(String message) =>
@@ -8,13 +9,36 @@ class ApiErrorHelper {
   static bool isAuthenticationError(String message) =>
       message.equalsIgnoreCases(authentication);
 
+  static bool isGraphQlError(String message) =>
+      message.equalsIgnoreCases(graphql);
+
+  //////////////////////////////////////////////////////////////////////////////
+  //////////////////////////////////////////////////////////////////////////////
+  //////////////////////////////////////////////////////////////////////////////
+
+  static String errorMainMessage(GraphQLError error) => error.message;
+
   static String errorCategory(GraphQLError error) => error.extensions[category];
 
+  static Map<String, dynamic> errorValidation(dynamic extension) =>
+      extension[validation];
+
+  static Map<String, dynamic> errorAuthentication(dynamic extension) =>
+      extension[authentication];
+
   static String errorReason(dynamic extension) => extension[reason];
+
+  //////////////////////////////////////////////////////////////////////////////
+  //////////////////////////////////////////////////////////////////////////////
+  //////////////////////////////////////////////////////////////////////////////
 
   static GraphQLError mapToModel(dynamic raw) {
     return GraphQLError.fromJSON(raw);
   }
+
+  //////////////////////////////////////////////////////////////////////////////
+  //////////////////////////////////////////////////////////////////////////////
+  //////////////////////////////////////////////////////////////////////////////
 
   static String handle(QueryResult result) {
     if (result.exception.clientException is NetworkException) {
