+++ lib/mmd_ecommerce.dart
@@ -15,8 +15,8 @@ export './generatedql/user/graphql_api.dart' hide AuthPayload, User;
 export './generatedql/address/graphql_api.dart';
 export './mmd_ecommerce.dart';
 export './mmd_ecommerce_fl_lib.dart';
-export './apis/api_keys.dart';
-export './apis/api_error_helper.dart';
+export './apis/base/api_keys.dart';
+export 'apis/error/api_error_helper.dart';
 
 /// we hide the AuthPayload from "./generatedql/user/graphql_api.dart"
 /// because we use the same name with "./generatedql/auth/graphql_api.dart"
